+++ platform/src/main/java/org/stellar/anchor/platform/callback/RestRateIntegration.java
@@ -7,7 +7,6 @@ import com.google.gson.Gson;
 import java.lang.reflect.Type;
 import java.net.URI;
 import java.net.URISyntaxException;
-import java.util.Calendar;
 import java.util.Map;
 import okhttp3.HttpUrl;
 import okhttp3.HttpUrl.Builder;
@@ -20,8 +19,6 @@ import org.stellar.anchor.api.callback.GetRateResponse;
 import org.stellar.anchor.api.callback.RateIntegration;
 import org.stellar.anchor.api.exception.AnchorException;
 import org.stellar.anchor.api.exception.ServerErrorException;
-import org.stellar.anchor.sep10.JwtService;
-import org.stellar.anchor.sep10.JwtToken;
 import org.stellar.anchor.util.Log;
 import shadow.com.google.common.reflect.TypeToken;
 
@@ -29,17 +26,10 @@ public class RestRateIntegration implements RateIntegration {
   private final String anchorEndpoint;
   private final OkHttpClient httpClient;
   private final Gson gson;
-  private final JwtService jwtService;
-  private final long jwtExpirationMilliseconds;
-  private final String issuerUrl;
+  private final AuthHelper authHelper;
 
   public RestRateIntegration(
-      String anchorEndpoint,
-      OkHttpClient httpClient,
-      JwtService jwtService,
-      String issuerUrl,
-      long jwtExpirationMilliseconds,
-      Gson gson) {
+      String anchorEndpoint, OkHttpClient httpClient, AuthHelper authHelper, Gson gson) {
     try {
       new URI(anchorEndpoint);
     } catch (URISyntaxException e) {
@@ -48,9 +38,7 @@ public class RestRateIntegration implements RateIntegration {
 
     this.anchorEndpoint = anchorEndpoint;
     this.httpClient = httpClient;
-    this.jwtService = jwtService;
-    this.issuerUrl = issuerUrl;
-    this.jwtExpirationMilliseconds = jwtExpirationMilliseconds;
+    this.authHelper = authHelper;
     this.gson = gson;
   }
 
@@ -66,12 +54,13 @@ public class RestRateIntegration implements RateIntegration {
           }
         });
     HttpUrl url = urlBuilder.build();
+    String authHeader = "Bearer " + authHelper.createJwtToken();
 
     Request httpRequest =
         new Request.Builder()
             .url(url)
             .header("Content-Type", "application/json")
-            .header("Authorization", "Bearer " + createJwtToken())
+            .header("Authorization", authHeader)
             .get()
             .build();
     Response response = call(httpClient, httpRequest);
@@ -109,12 +98,5 @@ public class RestRateIntegration implements RateIntegration {
       }
     }
     return getRateResponse;
-  }
-
-  String createJwtToken() {
-    long issuedAt = Calendar.getInstance().getTimeInMillis() / 1000L;
-    long expirationTime = issuedAt + (jwtExpirationMilliseconds / 1000L);
-    JwtToken token = JwtToken.of(issuerUrl, issuedAt, expirationTime);
-    return jwtService.encode(token);
   }
 }
